==> Exploracion/01_data_preparation_perfil_clientes.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np

df_perfil_clientes = pd.read_csv('df_final_demo.txt')

df_perfil_clientes.shape

df_perfil_clientes.columns

### Diccionario de variables – df_final_demo

- client_id: identificador único del cliente
- clnt_tenure_yr: antigüedad del cliente en años
- clnt_tenure_mnth: antigüedad del cliente en meses
- clnt_age: edad del cliente
- gendr: género del cliente
- num_accts: número de cuentas asociadas al cliente
- bal: balance total del cliente
- calls_6_mnth: llamadas al call center en los últimos 6 meses
- logons_6_mnth: número de accesos digitales en los últimos 6 meses


#Renombramos columnas para mejorar la comprension de lo que implica cada una de ellas y trabajar con nombres mas cortos y faciles de escribir

df_perfil_clientes = df_perfil_clientes.rename(columns={
    'clnt_tenure_yr': 'tenure_years',
    'clnt_tenure_mnth': 'tenure_months',
    'clnt_age': 'age',
    'gendr': 'gender',
    'num_accts': 'num_accounts',
    'bal': 'balance',
    'calls_6_mnth': 'calls_last_6_months',
    'logons_6_mnth': 'logins_last_6_months'
})


df_perfil_clientes.columns


df_perfil_clientes.head(10)


#Eliminamos la columna tenue_years, ya que es redundante con la columna tenure_months, que creemos que servirá más al análisis, siendo más intuitivo y práctico trabajar con meses que con años, ya que la mayoría de los clientes tienen una antigüedad menor a 5 años, lo que se traduce en menos de 60 meses, por lo que la columna tenure_months nos dará una mejor perspectiva de la antigüedad de los clientes.

df_perfil_clientes = df_perfil_clientes.drop(columns=['tenure_years'])


df_perfil_clientes.columns

#Verificación de la variable 'gender' para asegurarnos de que no haya valores atípicos o errores de tipeo, ya que es una variable categórica que solo debería contener

df_perfil_clientes['gender'].value_counts(dropna=False)


df_perfil_clientes['gender'] = (
    df_perfil_clientes['gender']
    .str.strip()
    .str.upper()
    .replace({
        'M': 'Male',
        'F': 'Female',
        'U': 'Unknown',
        'NaN' : 'Unknown',
        ' NaN' : 'Unknown',
        'NaN ' : 'Unknown',
        'X': 'Unknown'
    })
)


df_perfil_clientes['gender'] = df_perfil_clientes['gender'].fillna('UNKNOWN')
df_perfil_clientes['gender'].value_counts(dropna=False)


#Análisis de la variable 'age' para detectar posibles valores atípicos o errores de tipeo, ya que es una variable numérica que debería contener valores dentro de un rango razonable para la edad de los clientes.

df_perfil_clientes['age'].describe()


df_perfil_clientes['age'].isna().sum()


df_perfil_clientes.sort_values('age').head(10)


df_perfil_clientes.sort_values('age', ascending=False).head(10) 

#Se observa casos que hay menores de edad, tratándose de un grupo de inversión, con lo cual se procede a categorizar a esos clientes como "<18" para diferenciarlos del resto de los clientes, ya que no es posible que tengan una edad negativa o cero, lo cual es un error de tipeo o un valor atípico.

df_perfil_clientes.loc[
    df_perfil_clientes['age'] < 18,
    'age'
] = pd.NA


df_perfil_clientes['age_group'] = pd.cut(
    df_perfil_clientes['age'],
    bins=[18, 30, 45, 60, 75, 100],
    labels=['18–29', '30–44', '45–59', '60–74', '75+'],
    right=False
)


df_perfil_clientes['age_group'].value_counts(dropna=False)


df_perfil_clientes = df_perfil_clientes.rename(columns={
    'age': 'age_group'
})


df_perfil_clientes.columns


#Luego de analizar y modificar el nombre de la variable 'age' a 'age_group', procecemos a analizar 'balance' para detectar posibles valores atípicos o errores de tipeo, ya que es una variable numérica que debería contener valores dentro de un rango razonable para el saldo de los clientes.

df_perfil_clientes['balance'].isna().sum()


df_perfil_clientes['balance'].describe()


df_perfil_clientes.sort_values('balance').head(10)


df_perfil_clientes.sort_values('balance', ascending=False).head(10)


#Verificamos que haya clientes con saldo negativo o sin activos.

(df_perfil_clientes['balance'] < 0).sum()


(df_perfil_clientes['balance'] == 0).sum()


#“Dado que la variable balance presenta una distribución muy asimétrica, se aplicó una transformación logarítmica para facilitar la visualización y comparación entre grupos, manteniendo intacta la información relativa.”

df_perfil_clientes['balance_log'] = np.log1p(df_perfil_clientes['balance'])


df_perfil_clientes['balance'].describe()
df_perfil_clientes['balance'].describe().apply(lambda x: f"{x:,.0f}")


#Análisis de la variable 'logins_last_6_months' para detectar posibles valores atípicos o errores de tipeo, ya que es una variable numérica que debería contener valores dentro de un rango razonable para la cantidad de logins en los últimos 6 meses.

df_perfil_clientes['logins_last_6_months'].describe()


df_perfil_clientes['logins_last_6_months'].isna().sum()


df_perfil_clientes.sort_values('logins_last_6_months').head(10)


df_perfil_clientes.sort_values('logins_last_6_months', ascending=False).head(10)


df_perfil_clientes['calls_last_6_months'].isna().sum()


df_perfil_clientes.sort_values('calls_last_6_months').head(10)


df_perfil_clientes.sort_values('calls_last_6_months', ascending=False).head(10)


df_perfil_clientes = df_perfil_clientes.rename(columns={
    'logins_last_6_months': 'logins_6m',
    'calls_last_6_months': 'calls_6m'
})
df_perfil_clientes.columns

df_perfil_clientes[['logins_6m', 'calls_6m']] = (
    df_perfil_clientes[['logins_6m', 'calls_6m']]
    .astype('Int64')
)
df_perfil_clientes.dtypes   

df_perfil_clientes.info()


df_perfil_clientes = df_perfil_clientes.loc[:,~df_perfil_clientes.columns.duplicated()]


df_perfil_clientes['age_group'] = df_perfil_clientes['age_group'].astype('category')


df_perfil_clientes.info()


#Reorganizamos las columnas, según prioridad de análisis, colocando primero las variables demográficas, luego las variables de comportamiento y finalmente las variables financieras, para facilitar la interpretación y el análisis de los datos.

df_perfil_clientes = df_perfil_clientes[
    [
        'client_id',
        'age_group',
        'gender',
        'tenure_months',
        'num_accounts',
        'balance',
        'balance_log',
        'logins_6m',
        'calls_6m'
    ]
]


age_group: grupo de edad del cliente

gender: género normalizado

tenure_months: antigüedad del cliente en meses

num_accounts: número de cuentas activas

balance: saldo total del cliente

balance_log: saldo transformado logarítmicamente

logins_6m: número de logins en los últimos 6 meses

calls_6m: número de llamadas al soporte en los últimos 6 meses

## df_perfil_clientes – Estado final

Este dataframe contiene información demográfica y de comportamiento agregado de los clientes.
Las variables han sido limpiadas, tipadas y estandarizadas para su posterior integración con los
datasets de experimento y comportamiento web.

Clave de unión: `client_id`


df_perfil_clientes.to_csv("df_perfil_clientes_final.csv", index=False)
