==> Exploracion/02_data_preparation_final_experiment.ipynb <==
import pandas as pd



## Carga de Datos
Cargamos los datasets necesarios: `df_final_experiment_clients.txt`

df = pd.read_csv('df_final_experiment_clients.txt')


print(f"Experiment Clients: {df.shape}")


##  Renombrado y Tipos de Datos
Renombramos `Variation` a `experiment_group` para mayor claridad y verificamos los tipos de datos.

# Renombrar columna
df.rename(columns={'Variation': 'experiment_group'}, inplace=True)

# Verificación de tipos de datos
print("Tipos de datos actuales:")
print(df.dtypes)



## Análisis Exploratorio

# Muestra de datos
display(df.head())

# Análisis de la columna 'experiment_group' (Categorías)
print("Valores únicos en la columna experiment_group:")
print(df['experiment_group'].value_counts(dropna=False))

# Conteo de Nulos en experiment_group
nulls = df['experiment_group'].isnull().sum()
print(f"Nulos en experiment_group: {nulls} ({df['experiment_group'].isnull().mean() * 100:.2f}%) ")

# Conteo de Duplicados
duplicados = df.duplicated().sum()
print(f"Filas totalmente duplicadas: {duplicados}")

# Duplicados por Client ID
duplicados_id = df['client_id'].duplicated().sum()
print(f"Clientes repetidos: {duplicados_id}")

- Nulos Críticos: Identificados ***20,109*** usuarios 
- Desbalanceo de Grupos:
   Test: ***26,968*** usuarios.
   Control: ***23,532*** usuarios.
- Renombrado de columna Variation a experiment_group.

df.to_csv("df_experiment_final.csv", index=False)
